Move markdown galaxy section components to subdirectory
