+++ client/src/components/Markdown/Sections/Elements/HistoryDatasetAsImage.vue
