+++ client/src/components/Markdown/Sections/Elements/JobMetrics.vue
