+++ client/src/components/Markdown/Sections/Elements/Workflow/WorkflowImage.vue
